refactor(services): reorganize ShortcutTriggerState type and update imports across services for consistency
