+++ apps/app/src/lib/commands.ts
@@ -7,7 +7,7 @@ import {
 	deliverTransformedText,
 } from './deliverTextToUser';
 import { CommandOrAlt, CommandOrControl } from '$lib/constants/keyboard';
-import type { ShortcutTriggerState } from './services/shortcuts/shortcut-trigger-state';
+import type { ShortcutTriggerState } from './services/_shortcut-trigger-state';
 
 const stopManualRecording = async () => {
 	const toastId = nanoid();
