+++ apps/app/src/lib/services/_shortcut-trigger-state.ts
