+++ apps/app/src/lib/services/transcription/elevenlabs.ts
@@ -1,7 +1,7 @@
 import type { ElevenLabsModel } from '$lib/constants/transcription';
 import type { WhisperingError } from '$lib/result';
 import type { Settings } from '$lib/settings';
-import { Err, type Result } from '@epicenterhq/result';
+import { Err, Ok, type Result } from '@epicenterhq/result';
 import { ElevenLabsClient } from 'elevenlabs';
 
 export function createElevenLabsTranscriptionService() {
